优化多单元格复制操作，接触对其他区域的影响
