+++ version.txt
@@ -6,7 +6,7 @@ VSVersionInfo(
   ffi=FixedFileInfo(
     # filevers and prodvers should be always a tuple with four items: (1, 2, 3, 4)
     # Set not needed items to zero 0.
-    filevers=(1, 2, 32, 2024),
+    filevers=(1, 2, 33, 2024),
     # Contains a bitmask that specifies the valid bits 'flags'r
     mask=0x3f,
     # Contains a bitmask that specifies the Boolean attributes of the file.
@@ -34,7 +34,7 @@ VSVersionInfo(
         StringStruct('LegalCopyright', 'Copyright (C) 2022 kestiny'),
         StringStruct('OriginalFilename', 'DBcat.exe'),
         StringStruct('ProductName', 'DBcat'),
-        StringStruct('ProductVersion', '1.2.32.2024')])
+        StringStruct('ProductVersion', '1.2.33.2024')])
       ]), 
     VarFileInfo([VarStruct('Translation', [1033, 1200])])
   ]
